Some early fixes.

+++ apps/wget/src/net/crud/Wget/Wget.java
@@ -48,20 +48,20 @@ public class Wget extends Activity {
 				return true;
 			}
 		});		
-		Button about = (Button) findViewById(R.id.HelpButton);
+		Button help = (Button) findViewById(R.id.HelpButton);
 		final Context context = this.getApplicationContext();
-		about.setOnClickListener(new View.OnClickListener() {
+		help.setOnClickListener(new View.OnClickListener() {
 			public void onClick(View v) {
 				startActivity(new Intent(
                         Wget.this, WgetHelp.class));
 
 			}
 		});
-		Button help = (Button) findViewById(R.id.AboutButton);
-		help.setOnClickListener(new View.OnClickListener() {
+		Button about = (Button) findViewById(R.id.AboutButton);
+		about.setOnClickListener(new View.OnClickListener() {
 			public void onClick(View v) {
-				
-				// ...
+				startActivity(new Intent(
+                        Wget.this, WgetAbout.class));
 			}
 		});
 		
@@ -70,7 +70,6 @@ public class Wget extends Activity {
 	public void runWget() {
 		String base = this.getApplicationContext().getFilesDir().getParent();
 		String wget = base + "/wget";
-		this.copyBinary("wget");
 
 		String options = ((EditText) findViewById(R.id.OptionsField)).getText()
 				.toString();
@@ -83,35 +82,39 @@ public class Wget extends Activity {
 
 		TextView tv = (TextView) findViewById(R.id.output);
 		ScrollView sv = (ScrollView) findViewById(R.id.scrollview);
+
+		try {
+			this.copyBinary("wget");
+		} catch (IOException e) {
+		  	tv.setText(e.toString());
+		}
 		
 		new WgetTask(tv, sv).execute(wget + " " + args);
 	}
 
-	// Brazenly stolen from android-wifi-tether
-	// XXX Check if binary is already present and do nothing
-	private String copyBinary(String filename) {
+	private void copyBinary(String filename) throws IOException {
 		String base = this.getApplicationContext().getFilesDir().getParent();
 		String outFileName = base + "/" + filename;
 		File outFile = new File(outFileName);
+		if (!new File(outFileName).exists()) {
+			try {
+				Log.d("wget", "Extracting " + filename + " to " + outFileName);
+				InputStream is = this.getAssets().open(filename);
+				byte buf[] = new byte[1024];
+				int len;
+				OutputStream out = new FileOutputStream(outFile);
+				while ((len = is.read(buf)) > 0) {
+					out.write(buf, 0, len);
+				}
+				out.close();
+				is.close();
+				String[] cmd = { "/system/bin/chmod", "0755", outFileName };
 
-		Log.d("wget", "Copying file '" + filename + "' ...");
-		try {
-			InputStream is = this.getAssets().open(filename);
-			byte buf[] = new byte[1024];
-			int len;
-			OutputStream out = new FileOutputStream(outFile);
-			while ((len = is.read(buf)) > 0) {
-				out.write(buf, 0, len);
+				Process proc = Runtime.getRuntime().exec(cmd);
+			} catch (IOException e) {
+				throw e;
 			}
-			out.close();
-			is.close();
-			String[] cmd = { "/system/bin/chmod", "0755", outFileName };
-
-			Process proc = Runtime.getRuntime().exec(cmd);
-		} catch (IOException e) {
-			return "Couldn't install file - " + filename + "!";
 		}
-		return null;
 	}
 
 }
